fix: Clean up the code
